Revert "fix: refactor package name"
This reverts commit 53494bcb77cd97381a07f9cca11862fcc0a6a2e4.
