--- src/main/java/org/dspappas/AmbiguitiesService.java
@@ -1,4 +1,0 @@
-package org.dspappas;
-
-public class AmbiguitiesService {
-}
