+++ src/main/java/org/example/GreekNumberValidation.java
@@ -1,4 +1,4 @@
-package org.dspappas;
+package org.example;
 
 import org.jetbrains.annotations.NotNull;
