+++ src/main/java/org/example/Main.java
@@ -1,4 +1,4 @@
-package org.dspappas;
+package org.example;
 
 public class Main {
     public static void main(String[] args) {
